added support for commits made to CHANGELOG.md and missing changelog entries
You can't go back and re-insert .changelog_entries/ files
for old commits on a project that was directly editing it.
You'd be rewriting history

But, we can write a tool that parses those old commits,
and generates ChangelogEntry objects for them. Then adds those
in a NEW commit with a name that has the treeish of the old commit
plus .json.

If that has been done, we now support it.