[input] fix: readline related breakages
Something appears to have changed / broken in
Readline
it was screwing up the last line of the change
time question.

So, now i'm just using STDOUT and then passing
readline an empty string.

UX is slightly worse in that now you're entering
on the next line, but at least it doesn't look
bugged.